+++ lib/loxxy/back_end/lox_class.rb
@@ -1,10 +1,11 @@
 # frozen_string_literal: true
 
 require_relative '../datatype/all_datatypes'
+require_relative 'lox_instance'
 
 module Loxxy
   module BackEnd
-    # Representation of a Lox class.
+    # Runtime representation of a Lox class.
     class LoxClass
       # @return [String] The name of the class
       attr_reader :name
@@ -25,6 +26,15 @@ module Loxxy
         stack.push self
       end
 
+      def arity
+        0
+      end
+
+      def call(engine, _visitor)
+        instance = LoxInstance.new(self, engine)
+        engine.stack.push(instance)
+      end
+
       # Logical negation.
       # As a function is a truthy thing, its negation is thus false.
       # @return [Datatype::False]
@@ -32,7 +42,7 @@ module Loxxy
         Datatype::False.instance
       end
 
-      # Text representation of a Lox function
+      # Text representation of a Lox class
       def to_str
         name
       end